requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
openai==1.51.0
apscheduler==3.10.4
python-telegram-bot==21.6
httpx==0.27.2
beautifulsoup4==4.12.3
aiofiles==24.1.0
jinja2==3.1.4
python-multipart==0.0.12
sse-starlette==2.1.3
tenacity==9.0.0
rich==13.8.1
click==8.1.7
